Move element render functions into a separate root folder
