+++ src/render/BarLine.ts
@@ -1,5 +1,5 @@
-import layout from "../../../layout";
-import { Application } from "../../state/Application";
+import layout from "../layout";
+import { Application } from "../ui/state/Application";
 
 export const BarLine = (_application: Application, context: CanvasRenderingContext2D, element: layout.BarLine) => {
   context.lineWidth = element.strokeSize;
