+++ src/render/Dot.ts
@@ -1,5 +1,5 @@
-import * as layout from "../../../layout";
-import { Application } from "../../state/Application";
+import layout from "../layout";
+import { Application } from "../ui/state/Application";
 
 export const Dot = (_application: Application, context: CanvasRenderingContext2D, element: layout.Dot) => {
   const w = element.box.width;
