+++ src/render/Line.ts
@@ -1,6 +1,5 @@
-import * as layout from "../../../layout";
-import { LINE_STROKE_WIDTH } from "../../../layout";
-import { Application } from "../../state/Application";
+import layout, { LINE_STROKE_WIDTH } from "../layout";
+import { Application } from "../ui/state/Application";
 
 export const Line = (_application: Application, context: CanvasRenderingContext2D, element: layout.Line) => {
   context.lineWidth = LINE_STROKE_WIDTH;
