+++ src/render/Part.ts
@@ -1,5 +1,5 @@
-import layout from "../../../layout";
-import { Application } from "../../state/Application";
+import layout from "../layout";
+import { Application } from "../ui/state/Application";
 
 export const Part = (_application: Application, _context: CanvasRenderingContext2D, _part: layout.Part) => {
   // Nothing to do right now!
